move mip generation
